fix: Move _get_tfm_regression_targets() to psa_builder.py
Move _get_tfm_regression_targets() to psa_builder.py so targets
can be shared across the scripts.
