+++ psa_builder.py
@@ -25,6 +25,14 @@ import logging
 import requests
 from zipfile import ZipFile
 
+try:
+    import yaml
+except ImportError as e:
+    print(str(e) + " To install it, type:")
+    print("python3 -m pip install PyYAML")
+    exit(1)
+
+
 upstream_tfm = "https://git.trustedfirmware.org/TF-M/trusted-firmware-m.git"
 mbed_tfm = "https://github.com/ARMmbed/trusted-firmware-m.git"
 
@@ -76,6 +84,8 @@ TF_M_RELATIVE_PATH = (
 sys.path.insert(0, mbed_path)
 TF_M_BUILD_DIR = join(mbed_path, TF_M_RELATIVE_PATH, "TARGET_IGNORE")
 POPEN_INSTANCE = None
+
+from tools.targets import Target, TARGET_MAP, TARGET_NAMES
 
 
 def are_dependencies_installed():
@@ -224,3 +234,34 @@ def exit_gracefully(signum, frame):
         pass
 
     sys.exit(0)
+
+
+def get_tfm_secure_targets():
+    """
+    Creates a list of TF-M secure targets from Mbed OS targets.json.
+
+    :return: List of TF-M secure targets.
+    """
+    return [str(t) for t in TARGET_NAMES if Target.get_target(t).is_TFM_target]
+
+
+def get_tfm_regression_targets():
+    """
+    Creates a list of TF-M regression tests supported targets
+    This parses the yaml file for supported target names and compares them
+    with TF-M targets supported in Mbed OS
+
+    :return: List of supported TF-M regression targets.
+    """
+    with open(join(dirname(__file__), "tfm_ns_import.yaml")) as ns_import:
+        yaml_data = yaml.safe_load(ns_import)
+        mbed_os_data = yaml_data["mbed-os"]
+        tfm_regression_data = yaml_data["tf-m-regression"]
+
+        regression_targets = list(
+            set(get_tfm_secure_targets())
+            & set(mbed_os_data)
+            & set(tfm_regression_data)
+        )
+
+        return regression_targets
