+++ test_psa_target.py
@@ -272,7 +272,7 @@ def _get_parser():
         "--mcu",
         help="Build for the given MCU",
         required=True,
-        choices=["ARM_MUSCA_B1", "ARM_MUSCA_S1"],
+        choices=get_tfm_regression_targets(),
         default=None,
     )
